修正数据类型, 添加示例

+++ src/main/kotlin/com/hiczp/bilibili/api/app/model/Mine.kt
@@ -30,7 +30,7 @@ data class Mine(
             @SerializedName("level")
             var level: Int, // 5
             @SerializedName("mid")
-            var mid: Int, // 2866663
+            var mid: Long, // 2866663
             @SerializedName("name")
             var name: String, // hyx5020
             @SerializedName("new_followers")
